portofolio/lontara-usle-project/backend/requirements.txt
flask
flask-cors
earthengine-api
python-dotenv